chore: Separate queries by post type.
