+++ src/app/page.tsx
@@ -3,7 +3,7 @@ import { EquipoSection } from "@/components/sections/EquipoSection";
 import { ParaQuienSection } from "@/components/sections/ParaQuienSection";
 import { PortfolioShowCaseSection } from "@/components/sections/PortfolioShowCaseSection";
 import { ProductosSection } from "@/components/sections/ProductosSection";
-import { GET_ALL_PRODUCTS, GET_ALL_SLIDES } from "@/graphql/queries";
+import { GET_ALL_PRODUCTS, GET_ALL_SLIDES } from "@/graphql";
 import { client } from "@/lib/apollo";
 import { Product, SlideResponse } from "@/lib/types";
