+++ src/graphql/index.ts
@@ -1,0 +1,3 @@
+export * from "./portfolio/queries";
+export * from "./slides/queries";
+export * from "./products/queries";
