+++ src/graphql/slides/queries.ts
@@ -1,0 +1,25 @@
+import { gql } from "@apollo/client";
+
+export const GET_ALL_SLIDES = gql`
+  query GetAllSlides {
+    slides {
+      nodes {
+        slidesFg {
+          caption
+          title
+          description
+          buttonText
+          buttonLink
+          image {
+            node {
+              altText
+              sourceUrl
+            }
+          }
+          order
+          active
+        }
+      }
+    }
+  }
+`;
